# faster semaphore group instantiation