--- a/packages/input-test-pcd/package.json
+++ b/packages/input-test-pcd/package.json
@@ -19,9 +19,9 @@
   },
   "dependencies": {
     "@pcd/pcd-types": "0.5.2",
-    "@semaphore-protocol/group": "^3.2.3",
-    "@semaphore-protocol/identity": "^3.2.3",
-    "@semaphore-protocol/proof": "^3.2.3",
+    "@semaphore-protocol/group": "^3.10.0",
+    "@semaphore-protocol/identity": "^3.10.0",
+    "@semaphore-protocol/proof": "^3.10.0",
     "json-bigint": "^1.0.0",
     "uuid": "^9.0.0"
   },