--- a/packages/passport-crypto/package.json
+++ b/packages/passport-crypto/package.json
@@ -21,7 +21,7 @@
     "@pcd/pcd-collection": "0.5.2",
     "@pcd/pcd-types": "0.5.2",
     "@pcd/semaphore-identity-pcd": "0.5.2",
-    "@semaphore-protocol/identity": "^3.2.3",
+    "@semaphore-protocol/identity": "^3.10.0",
     "libsodium-wrappers": "^0.7.11",
     "react": "^18.2.0"
   },