--- a/packages/passport-interface/package.json
+++ b/packages/passport-interface/package.json
@@ -23,9 +23,9 @@
     "@pcd/semaphore-group-pcd": "0.5.2",
     "@pcd/semaphore-identity-pcd": "0.5.2",
     "@pcd/semaphore-signature-pcd": "0.5.2",
-    "@semaphore-protocol/group": "^3.2.3",
-    "@semaphore-protocol/identity": "^3.2.3",
-    "@semaphore-protocol/proof": "^3.2.3",
+    "@semaphore-protocol/group": "^3.10.0",
+    "@semaphore-protocol/identity": "^3.10.0",
+    "@semaphore-protocol/proof": "^3.10.0",
     "react": "^18.2.0"
   },
   "devDependencies": {