--- a/packages/semaphore-identity-pcd/package.json
+++ b/packages/semaphore-identity-pcd/package.json
@@ -20,9 +20,9 @@
   "dependencies": {
     "@pcd/passport-ui": "0.5.2",
     "@pcd/pcd-types": "0.5.2",
-    "@semaphore-protocol/group": "^3.2.3",
-    "@semaphore-protocol/identity": "^3.2.3",
-    "@semaphore-protocol/proof": "^3.2.3",
+    "@semaphore-protocol/group": "^3.10.0",
+    "@semaphore-protocol/identity": "^3.10.0",
+    "@semaphore-protocol/proof": "^3.10.0",
     "json-bigint": "^1.0.0",
     "react": "^18.2.0",
     "styled-components": "^5.3.9",