closes https://github.com/proofcarryingdata/zupass/issues/279
closes https://github.com/proofcarryingdata/zupass/issues/263

## Summary

- tl;dr after incorporating the latest semaphore package with improvement described [here](https://github.com/semaphore-protocol/semaphore/issues/319), performance of semaphore group-related actions improved by factors between `5x` and `20x`
- Server-side instantiation of the semaphore groups went down from `6s` to an average of `300ms`. This one is important to make fast because we rebuild semaphore groups every time someone logs in. Currently, the time a user has to wait between submitting their email verification code and being logged in is around `16s`. Based on the improvement I am seeing locally, I expect logging in to take around `800ms` in production.
- Client-side proving of zuzalu group membership went down from `8.3s` to `1.7s` locally on my computer, which is appx `5x` faster. This operation is important to make fast because users have to make these proofs to interact with zupoll and zucast, sometimes once per action.

## Evidence
### Server-side semaphore group reloading
<img width="671" alt="Screenshot 2023-05-15 at 11 17 28 AM" src="https://github.com/proofcarryingdata/zupass/assets/2636237/cbfaa7d0-a33d-4de7-aa19-678fc46db1aa">

#### traces of reload before improvement
<img width="689" alt="Screenshot 2023-05-15 at 11 19 18 AM" src="https://github.com/proofcarryingdata/zupass/assets/2636237/d9e85cbd-09a4-4932-891e-0f0fda978b37">
<img width="685" alt="Screenshot 2023-05-15 at 11 19 02 AM" src="https://github.com/proofcarryingdata/zupass/assets/2636237/35f29555-21df-4035-93e0-e5677856b012">


#### tracess of reload after improvement
<img width="720" alt="Screenshot 2023-05-15 at 11 11 46 AM" src="https://github.com/proofcarryingdata/zupass/assets/2636237/99eeed7c-3cae-4942-874f-6d4a5b679a5b">
<img width="767" alt="Screenshot 2023-05-15 at 11 10 56 AM" src="https://github.com/proofcarryingdata/zupass/assets/2636237/82c6b737-1276-49d7-bd23-fd4666c2d73a">


### Client-side group membership proofs

#### performance before improvement
<img width="367" alt="Screenshot 2023-05-15 at 11 20 14 AM" src="https://github.com/proofcarryingdata/zupass/assets/2636237/75c9d31a-0101-48af-847d-055a0bcade8f">

#### performance after improvement
<img width="362" alt="Screenshot 2023-05-15 at 11 13 31 AM" src="https://github.com/proofcarryingdata/zupass/assets/2636237/bd894132-3803-4d51-bf9c-76c1453546af">